% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-filteringAnno.R
\name{annoCuration,genomicCompartments-method}
\alias{annoCuration,genomicCompartments-method}
\title{Curate Gene Annotations}
\usage{
annoCuration(theobject, geneannovec, nbcpu = 8)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{geneannovec}{A character vector containing the paths to gene
annotation files.}

\item{nbcpu}{Number of CPU cores to use for parallel processing
(default is 8).}
}
\value{
Returns the updated `genomicCompartments` object with curated gene
annotations.
}
\description{
This method curates gene annotations by performing several filtering steps
on the provided annotation data. It filters out non-canonical chromosomes,
retains known coding and non-coding genes, keeps genes with a length greater
than 2kb, removes duplicate annotations with the same coordinates, extends
the transcription start site (TSS) by 1kb upstream, and removes overlapping
annotations. The curated annotations are then set as the curated annotation
data of the `genomicCompartments` object.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Perform annotation curation
gc_obj <- annoCuration(gc_obj, geneannovec)
}

}
