% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-overlap.R
\name{bivalentPromoters,genomicCompartments-method}
\alias{bivalentPromoters,genomicCompartments-method}
\title{Identify Bivalent Promoters}
\usage{
bivalentPromoters(theobject, peakspathvec)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{peakspathvec}{A named vector specifying paths to the peak files of
H3K4me3 and H3K27me3.}
}
\value{
Returns the modified `genomicCompartments` object with bivalent
promoters assigned.
}
\description{
This method identifies bivalent promoters by creating a merged GRanges object
from peaks of H3K4me3 and H3K27me3 histone marks.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Identify bivalent promoters
bivalentPromoters(gc_obj, peakspathvec = c(H3K4me3 = "path/to/H3K4me3.bed",
H3K27me3 = "path/to/H3K27me3.bed"))
}

}
