% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildGR}
\alias{buildGR}
\title{Build GRanges Object from File}
\usage{
buildGR(currentpath)
}
\arguments{
\item{currentpath}{A character string representing the BED file path to read.
The file should contain tabular data with the following columns:
\itemize{
  \item \code{V1} - Chromosome names
  \item \code{V4} - Start positions
  \item \code{V5} - End positions
  \item \code{V9} - Feature names
  \item \code{V7} - Strand information
}}
}
\value{
A `GRanges` object with the specified genomic coordinates and
annotations.
}
\description{
This function reads a file containing genomic coordinates and constructs a
`GRanges` object.
}
\examples{
\dontrun{
# Example file path
path <- "path/to/genomic/coordinates/file.bed"
# Build GRanges object
gr <- buildGR(path)
}

}
