% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildIntervalsObject}
\alias{buildIntervalsObject}
\title{Build Genomic Intervals Object}
\usage{
buildIntervalsObject(peakspathvec, geneannos, includerepeats)
}
\arguments{
\item{peakspathvec}{A named vector of file paths to peak files for various
ChIP-seq and ATAC-seq data. Expected names include: "H3K27ac", "Ser5P",
"Ser2P", "H3K4me3", "H3K27me3", "ATACSeq", "Suz12", "RING1B", "H3K9me3".}

\item{geneannos}{A list of gene annotation files to be used for defining
non-redundant gene sets and other features.}

\item{includerepeats}{A logical value indicating whether to include repeat
annotations (LINE, LTR, SINE).}
}
\value{
An updated `genomeCompart` object with defined genomic features and,
optionally, repeat annotations.
}
\description{
This function constructs a `genomeCompart` object, defines various genomic
features, and optionally includes repeat annotations.
}
\details{
This function performs the following steps:
\itemize{
  \item Builds the initial `genomeCompart` object.
  \item Defines a non-redundant set of gencode genes by removing overlapping
genes, genes closer than 1kb to each other, and genes shorter than 2kb.
  \item Identifies active promoters using H3K27ac at TSS±1Kb.
  \item Identifies promoters with Ser5P (initiation) at TSS±1Kb.
  \item Identifies elongating genes by overlapping Ser2P from TSS+1kb to TES.
  \item Defines termination sites at TES+1kb.
  \item Identifies bivalent promoters using H3K4me3 and H3K27me3 at TSS±1kb.
  \item Defines active enhancers using H3K27ac, H3K4me1, and ATAC-seq,
ensuring no overlap with gene annotations.
  \item Defines poised enhancers using H3K27me3, H3K4me1, and Suz12,
ensuring no overlap with gene annotations.
  \item Defines polycomb domains using Suz12 and RING1B overlaps.
  \item Defines heterochromatin using H3K9me3.
  \item Optionally includes repeat annotations (SINE, LINE, LTR) if
`includerepeats` is TRUE.
}
}
\examples{
\dontrun{
peakspathvec <- c(H3K27ac = "path/to/H3K27ac.gff",
     Ser5P = "path/to/Ser5P.gff", Ser2P = "path/to/Ser2P.gff",
     H3K4me3 = "path/to/H3K4me3.gff", H3K27me3 = "path/to/H3K27me3.gff",
     ATACSeq = "path/to/ATACSeq.gff", Suz12 = "path/to/Suz12.gff",
     RING1B = "path/to/RING1B.gff", H3K9me3 = "path/to/H3K9me3.gff")
geneannos <- list("path/to/gencode.gff", "path/to/refGene.gff",
"path/to/refseq.gff")
includerepeats <- FALSE
gc <- buildIntervalsObject(peakspathvec, geneannos, includerepeats)
}

}
