% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{complexUpsetDiagram,list-method}
\alias{complexUpsetDiagram,list-method}
\title{Generate Complex Upset Diagram}
\usage{
complexUpsetDiagram(theobject, includerepeats, outfold)
}
\arguments{
\item{theobject}{A list of objects of class `genomicCompartments`.}

\item{outfold}{Character string specifying the output folder path for saving
the upset diagram and boxplot.}

\item{includerepeats}{Logical indicating whether to include repeats in the
analysis. Default is FALSE.}

\item{includeenhancers}{Logical indicating whether to include enhancers in
the analysis. Default is FALSE.}

\item{plotglclevels}{Logical indicating if the violin plot of the values of
the glc peaks should be plotted on top. Default is FALSE.}

\item{minsize}{Minimum number of overlap for a category to be indicated on
the plot. Default is 10.}
}
\value{
No return value. The function generates and saves a complex upset
plot in the specified output folder.
}
\description{
This method generates a plot showing the upset diagram of the overlaps of
GlcNAc peaks with different genomic compartments and the corresponding
average binding values for each peak.
}
\examples{
\dontrun{
# Create a list of genomicCompartments objects
gc_list <- list(gc_obj1, gc_obj2, gc_obj3)
# Generate complex upset diagram
complexUpsetDiagram(gc_list, includerepeats = FALSE,
 outfold = "path/to/output_folder")
}

}
