% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{.plotUpset}
\alias{.plotUpset}
\title{Generate UpSet Diagram}
\usage{
upsetDiagram(theobject, outputfolder, includerepeats, plotupset = FALSE)
}
\arguments{
\item{outputfolder}{Character string specifying the output folder where the
UpSet diagram will be saved.}

\item{plotupset}{Logical indicating whether to plot on device the upset plot
(`TRUE`) or not (`FALSE`).}

\item{theobject}{An object of class `genomicCompartments`.}

\item{includerepeats}{Logical indicating whether to include repeats in the
analysis.}
}
\value{
Does not return anything but saves the plot to outpufolder.
}
\description{
This method generates an UpSet diagram based on the overlap matrix of glucnac
peaks and genomic compartments.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Generate UpSet diagram
upsetDiagram(gc_obj, outputfolder = "results/upset_plots",
includerepeats = TRUE, plotupset = TRUE)
}

}
