% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{.violinPlot}
\alias{.violinPlot}
\title{Generate Violin Plot for Gene Expression}
\usage{
violinplotExpression(theobject, outfold, includerep, rnacounts, refseqpath,
 geneidtab, countsannotype, plotviolin = FALSE)
}
\arguments{
\item{plotviolin}{Logical indicating whether to plot on current device
(default is FALSE).}

\item{theobject}{An object of class `genomicCompartments`.}

\item{outfold}{Character string specifying the output folder path for saving
results.}

\item{includerep}{Logical indicating whether to include repeats in the
analysis.}

\item{rnacounts}{Character string specifying the path to the RNA-seq count
table.}

\item{refseqpath}{Character string specifying the path to the original
RefSeq annotation file.}

\item{geneidtab}{Table containing genes ID.}

\item{countsannotype}{Character string specifying the type of counts (e.g.,
entrez or not) used for expression values.}
}
\value{
Save the violin plot to outfold.
}
\description{
This method generates violin plots for gene expression across genomic
compartments.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Generate violin plots for gene expression
gc_obj <- violinplotExpression(gc_obj, outfold = "path/to/output_folder",
 includerep = TRUE, rnacounts = "path/to/rna_counts",
 refseqpath = "path/to/refseq_annotation",
 geneidtab = geneidtab, countsannotype = "entrez", plotviolin = TRUE)
}

}
