% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-output.R
\name{.writeCoord}
\alias{.writeCoord}
\title{Output Glc Peaks Coordinates per Compartment}
\usage{
outputGlcPeaksCoordPerCompartment(theobject, outputfolder, glcpeakspath,
                       includerepeats)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{outputfolder}{Path to the folder where output files will be written.}

\item{glcpeakspath}{Path to the GFF file containing glucnac peaks
coordinates.}

\item{includerepeats}{Logical indicating whether to include repeat regions
in the analysis.}
}
\value{
Returns nothing explicitly; outputs GFF and BED files to the
specified folder.
}
\description{
This method calculates the overlap of glucnac peaks with each compartment
defined in the `genomicCompartments` object and outputs the coordinates of
the overlapping peaks to specified output formats (GFF and BED).
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Output glucnac peaks coordinates per compartment
outputGlcPeaksCoordPerCompartment(gc_obj, outputfolder, glcpeakspath,
includerepeats = TRUE)
}

}
