% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-output.R
\name{extractCompCoordWithPeak,genomicCompartments-method}
\alias{extractCompCoordWithPeak,genomicCompartments-method}
\title{Extract Compartment Coordinates with Peaks}
\usage{
extractCompCoordWithPeak(theobject, outfold, includerep)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{outfold}{Path to the folder where output files will be written.}

\item{includerep}{Logical indicating whether to include repeat regions in
the analysis.}
}
\value{
Returns the modified `genomicCompartments` object with compartments
containing peaks.
}
\description{
This method performs overlap analysis between reference peaks and each
compartment defined in the `genomicCompartments` object. It extracts
compartments containing peaks and outputs their coordinates in GFF format.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Extract compartment coordinates with peaks
extractCompCoordWithPeak(gc_obj, outfold = "output_folder",
 includerep = TRUE)
}

}
