% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{genomeCompart}
\alias{genomeCompart}
\title{Create Genomic Compartments Object}
\usage{
genomeCompart(peakspathvec, geneannovec)
}
\arguments{
\item{peakspathvec}{Character vector containing the path to the peaks file.}

\item{geneannovec}{Character vector containing the paths to gene annotation
files.}
}
\value{
Returns an object of class `genomicCompartments` containing the
processed peaks and gene annotations.
}
\description{
This function reads glucnac peaks and gene annotations, converts them to
`GRanges` and `GRangesList` objects, respectively, and returns a
`genomicCompartments` object. It processes the input files by filtering out
unwanted annotations and chromosomes.
}
\examples{
\dontrun{
peakspathvec <- c("path/to/peaks.gff")
geneannovec <- c("path/to/geneanno1.gff", "path/to/geneanno2.gff")
genomic_compartments <- genomeCompart(peakspathvec, geneannovec)
}

}
