% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{matrixForUpset,genomicCompartments-method}
\alias{matrixForUpset,genomicCompartments-method}
\title{Create Matrix for UpSet Plot}
\usage{
matrixForUpset(theobject, includerepeats, glcpeakvalues)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{includerepeats}{Logical indicating whether to include repeats in the
analysis.}

\item{includeenhancers}{Logical indicating whether to include enhancers in
the analysis.}

\item{glcpeakvalues}{Numeric vector of glc peak values.}
}
\value{
Returns the modified `genomicCompartments` object with the regions
matrix assigned.
}
\description{
This method creates a matrix for generating an UpSet plot, indicating
overlaps between glc peaks and genomic compartments.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Create matrix for UpSet plot
matrixForUpset(gc_obj, includerepeats = FALSE, glcpeakvalues = c(1, 2, 3))
}

}
