% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-overlap.R
\name{overlapOnGenes,genomicCompartments-method}
\alias{overlapOnGenes,genomicCompartments-method}
\title{Overlap Peaks on Genes}
\usage{
overlapOnGenes(theobject, regionlabel, peaklabel = NA, peakpath = NA)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{regionlabel}{Character specifying the genomic region label ("TSS",
"GBTES", or "TES").}

\item{peaklabel}{Character specifying the peak label ("H3K27ac", "Ser5P",
or "Ser2P").}

\item{peakpath}{Path to the file containing peaks data. NA if regionlabel is
"TES".}
}
\value{
Returns the modified `genomicCompartments` object with assigned
compartments.
}
\description{
This method performs overlap analysis between peaks and curated annotations
at specified genomic regions (TSS or TES). It assigns overlapping peaks to
specific compartments based on the provided labels.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Perform overlap analysis on TSS with H3K27ac peaks
overlapOnGenes(gc_obj, regionlabel = "TSS", peaklabel = "H3K27ac",
peakpath = "path/to/peaks.bed")
}

}
