% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-overlap.R
\name{polycombsDomains,genomicCompartments-method}
\alias{polycombsDomains,genomicCompartments-method}
\title{Identify Polycomb Domains}
\usage{
polycombsDomains(theobject, peakspath1, peakspath2, label1, label2,
domainname)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{peakspath1}{Path to the peaks file of the first polycomb.}

\item{peakspath2}{Path to the peaks file of the second polycomb.}

\item{label1}{Label for the first polycomb.}

\item{label2}{Label for the second polycomb.}

\item{domainname}{Name or label for the identified Polycomb domains.}
}
\value{
Returns the modified `genomicCompartments` object with Polycomb
domains assigned.
}
\description{
This method identifies Polycomb domains based on the overlap of two
marks.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Identify Polycomb domains
polycombsDomains(gc_obj, peakspath1 = "path/to/SUZ12.gff",
 peakspath2 = "path/to/RING1B.gff",
 label1 = "SUZ12", label2 = "RING1B", domainname = "PolycombDomains")
}

}
