% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{retrieveGlcPeakVal,genomicCompartments-method}
\alias{retrieveGlcPeakVal,genomicCompartments-method}
\title{Retrieve Glc Peak Values}
\usage{
retrieveGlcPeakVal(theobject, includerepeats, bwpath)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{includerepeats}{Logical indicating whether to include repeats in the
analysis.}

\item{bwpath}{Character string specifying the path to the bigWig file
containing glc peak values.}
}
\value{
Returns the modified `genomicCompartments` object with glc peak
values assigned.
}
\description{
This method retrieves glc peak values associated with genomic
compartments.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Retrieve glc peak values
gc_obj <- retrieveGlcPeakVal(gc_obj, includerepeats = FALSE,
bwpath = "path/to/bigWig/files")
}

}
