% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setHeterochromatin<-,genomicCompartments-method}
\alias{setHeterochromatin<-,genomicCompartments-method}
\title{Allocates heterochromatic region coordinates}
\usage{
setHeterochromatin(obj) <- value
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{value}{GRanges object of heterochromatic region coordinates.}
}
\value{
The updated object of class `genomicCompartments` with the new
heterochromatin regions.
}
\description{
This method stores the heterochromatin region coordinates in an object of
class `genomicCompartments`.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- new("genomicCompartments")
# Define new heterochromatin regions
new_heterochromatin <- GRanges(seqnames = "chr1",
ranges = IRanges(start = 1, end = 1000000))
# Set the heterochromatin regions
setHeterochromatin(gc_obj) <- new_heterochromatin
}

}
