% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setLINE<-,genomicCompartments-method}
\alias{setLINE<-,genomicCompartments-method}
\title{setLINE}
\usage{
setLINE(obj) <- value
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{value}{GRanges object of LINE region coordinates.}
}
\value{
The updated object of class `genomicCompartments` with the new LINE
regions.
}
\description{
This method sets the LINE regions for an object of class
`genomicCompartments`.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- new("genomicCompartments")
# Define new LINE regions
new_LINE <- GRanges(seqnames = "chr1",
ranges = IRanges(start = 1, end = 1000000))
# Set the LINE regions
setLINE(gc_obj) <- new_LINE
}

}
