% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setLTR<-,genomicCompartments-method}
\alias{setLTR<-,genomicCompartments-method}
\title{Allocate LTR coordinates}
\usage{
setLTR(obj) <- value
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{value}{GRanges object of LTR region coordinates.}
}
\value{
The updated object of class `genomicCompartments` with the new LTR
regions.
}
\description{
This method stores the heterochromatin region coordinates in an object of
class `genomicCompartments`.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- new("genomicCompartments")
# Define new LTR regions
new_LTR <- GRanges(seqnames = "chr1",
ranges = IRanges(start = 1, end = 1000000))
# Set the LTR regions
setLTR(gc_obj) <- new_LINE
}

}
