% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setSINE<-,genomicCompartments-method}
\alias{setSINE<-,genomicCompartments-method}
\title{setSINE}
\usage{
setSINE(obj) <- value
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{value}{GRanges object of SINE region coordinates.}
}
\value{
The updated object of class `genomicCompartments` with the new SINE
regions.
}
\description{
This method sets the SINE regions for an object of class
`genomicCompartments`.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- new("genomicCompartments")
# Define new SINE regions
new_SINE <- GRanges(seqnames = "chr1",
ranges = IRanges(start = 1, end = 1000000))
# Set the SINE regions
setSINE(gc_obj) <- new_SINE
}

}
