% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{tryGetBM}
\alias{tryGetBM}
\title{Attempt to Retrieve Information from Ensembl Using biomaRt}
\usage{
tryGetBM(attributes, ensembl, values = NULL, filters = NULL)
}
\arguments{
\item{attributes}{Character vector specifying the attributes to retrieve.}

\item{ensembl}{`Mart` object representing the connection to the Ensembl
database.}

\item{values}{Optional. List or vector of values for the filter.}

\item{filters}{Optional. Character vector specifying the filters to apply.}
}
\value{
Returns a data frame with the retrieved information if the query is
successful.
}
\description{
This function attempts to retrieve information about genes from Ensembl using
the `biomaRt` package. It tries up to 5 times before giving up and raising an
error. The function supports specifying attributes, values, and filters for
the query.
}
\examples{
\dontrun{
ensembl <- tryUseMart(biomart = "ensembl", dataset = "hsapiens_gene_ensembl",
host = "https://www.ensembl.org")
attributes <- c("ensembl_gene_id", "external_gene_name", "chromosome_name",
"start_position", "end_position")
gene_info <- tryGetBM(attributes = attributes, ensembl = ensembl)
}

}
