% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{tryUseMart}
\alias{tryUseMart}
\title{Attempt to Connect to Ensembl Using biomaRt}
\usage{
tryUseMart(biomart = "ensembl", dataset, host, alternativemirror)
}
\arguments{
\item{biomart}{Character string specifying the BioMart database to use.
Default is "ensembl".}

\item{dataset}{Character string specifying the dataset to use within the
BioMart database.}

\item{host}{Character string specifying the host URL to connect to.}

\item{alternativemirror}{Logical value indicating whether to use an
alternative mirror. Default is `FALSE`.}
}
\value{
Returns a `Mart` object if the connection is successful.
}
\description{
This function attempts to establish a connection to the Ensembl database
using #' the `biomaRt` package. It tries up to 5 times before giving up and
raising an error. The function supports using an alternative mirror if
specified.
}
\examples{
\dontrun{
ensembl <- tryUseMart(biomart = "ensembl", dataset = "hsapiens_gene_ensembl",
host = "https://www.ensembl.org")
}

}
