% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{upsetDiagram,genomicCompartments-method}
\alias{upsetDiagram,genomicCompartments-method}
\title{Generate UpSet Diagram}
\usage{
upsetDiagram(theobject, outputfolder, includerepeats, plotupset = FALSE)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{outputfolder}{Character string specifying the output folder where the
UpSet diagram will be saved.}

\item{includerepeats}{Logical indicating whether to include repeats in the
analysis.}

\item{plotupset}{Logical indicating whether to plot on device the upset plot
(`TRUE`) or not (`FALSE`).}
}
\value{
Does not return anything but saves the plot to outpufolder.
}
\description{
This method generates an UpSet diagram based on the overlap matrix of glc
peaks and genomic compartments.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Generate UpSet diagram
upsetDiagram(gc_obj, outputfolder = "results/upset_plots",
includerepeats = FALSE, plotupset = TRUE)
}

}
