% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-overlap.R
\name{enhancer,genomicCompartments-method}
\alias{enhancer,genomicCompartments-method}
\title{Identify Enhancers}
\usage{
enhancer(theobject, peakspath1, peakspath2, peakspath3, label1, label2,
label3, statelabel)
}
\arguments{
\item{theobject}{An object of class `genomicCompartments`.}

\item{peakspath1}{Path to the peaks file of the first histone mark.}

\item{peakspath2}{Path to the peaks file of the second histone mark.}

\item{peakspath3}{Path to the peaks file of the third histone mark or
ATAC-seq.}

\item{label1}{Label for the first histone mark.}

\item{label2}{Label for the second histone mark.}

\item{label3}{Label for the third histone mark or ATAC-seq.}

\item{statelabel}{Label indicating the state of enhancers ('active' or
'poised').}
}
\value{
Returns the modified `genomicCompartments` object with enhancers
assigned.
}
\description{
This method identifies active or poisedd enhancers based on the overlap of
histone marks and ATAC-seq peaks.
}
\details{
Active enhancers are defined as the overlap of H3K27ac/H3K4me1/ATAC-seq.
Poised enhancers are defined as the overlap of H3K27me3/H3K4me1/PRC2.
They should not overlap with the combination of UCSC refGene, NCBI RefSeq,
and GENCODE VM25.
}
\examples{
\dontrun{
# Create a genomicCompartments object
gc_obj <- genomeCompart(peakspathvec, geneannovec)
# Identify active enhancers
enhancer(gc_obj, peakspath1 = "path/to/H3K27ac.gff",
 peakspath2 = "path/to/H3K4me1.gff",
 peakspath3 = "path/to/ATAC-seq.gff",
label1 = "H3K27ac", label2 = "H3K4me1", label3 = "ATAC-seq",
statelabel = "active")
}

}
