% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{genomeCompart}
\alias{genomeCompart}
\title{Create Genomic Compartments Object}
\usage{
genomeCompart(peakspathvec, geneannovec)
}
\arguments{
\item{peakspathvec}{Character vector containing the path to the peaks file.}

\item{geneannovec}{Character vector containing the paths to gene annotation
files.}
}
\value{
Returns an object of class `genomicCompartments` containing the
processed peaks and gene annotations.
}
\description{
This function is the constructor of the genomicCompartments object. It is
called by the function buildIntervalsObject.
}
\details{
peakspathvec is made of the GFF file containing glc peaks and of the GFF
files containing the ChIP-seq and ATAC-seq peak coordinates. The geneannovec
contains path to the gencode, refgene, and refseq gff files.
}
\examples{
\dontrun{
peakspathquery <- "/path/to/glc.gff"
peakspathcategoriesvec <- c(H3K27ac = "/path/to/H3K27ac.gff",
       H3K4me1 = "/path/to/H3K4me1.gff", H3K27me3 = "/path/to/H3K27me3.gff",
       H3K4me3 = "/path/to/H3K4me3.gff", Suz12 = NA, RING1B = NA,
       H3K9me3 = "/path/to/H3K9me3.gff", Ser5P = "/path/to/Ser5P.gff",
       Ser2P = "/path/to/Ser2P.gff", ATACSeq = "/path/to/ATACseq.gff")
peakspathvec <- c(peakspathquery, peakspathcategoriesvec)
geneannovec <- c(gencode = "/path/to/gencode.gff",
       refgene = "/path/to/refGeneUCSC.gff",
       refseq = "/path/to/refseq.gff")

genomic_compartments <- genomeCompart(peakspathvec, geneannovec)
}

}
\seealso{
buildIntervalsObject
}
